+++ libs/shared/src/lib/services/grid-data-formatter/grid-data-formatter.helper.ts
@@ -1,6 +1,5 @@
 import { get, isNil } from 'lodash';
 import { GridField } from '../../models/grid.model';
-import { ICON_EXTENSIONS } from '../../components/ui/core-grid/grid/grid.constants';
 
 /**
  * Returns field style from path.
@@ -30,32 +29,6 @@ export function getUrl(url: string): URL | null {
   } catch {
     return null;
   }
-}
-
-/**
- * Removes file extension from the file name
- *
- * @param name Name of the file with the extension
- * @returns String with the name of the file without the extension
- */
-export function removeFileExtension(name: string): string {
-  const fileExt = name.split('.').pop();
-  return fileExt && ICON_EXTENSIONS[fileExt]
-    ? name.slice(0, name.lastIndexOf(fileExt) - 1)
-    : name;
-}
-
-/**
- * Gets the kendo class icon for the file extension
- *
- * @param name Name of the file with the extension
- * @returns String with the name of the icon class
- */
-export function getFileIcon(name: string): string {
-  const fileExt = name.split('.').pop();
-  return fileExt && ICON_EXTENSIONS[fileExt]
-    ? ICON_EXTENSIONS[fileExt]
-    : 'k-i-file';
 }
 
 /**
